Improve determination of root permissions
